Adding configuration property GondolaArmLength_MM Allows a number to be entered which will be added to future (non-zero) L and R measurements entered at setup.
Changed GetArgsAsFloats to accept more than expectedCount parameters, as
already indicated by the "Expected at least, only saw" error value so that
setup can take 3 or more arguments.

Behavior is backward compatible to all uses prior to existance of the variable,
and a reminder is printed during the setup call if an existing non-zero
value for the property exists.

Hat tip to Brandon Dunson for pointing out that I had committed the wrong
version of the chart.fo Google Code fix, which was thus partially incomplete.